Refactor stories to not use mockedprovider
