+++ src/pages/version/taskDuration/TaskDurationTable.stories.tsx
@@ -1,17 +1,9 @@
-import { MockedProvider } from "@apollo/client/testing";
 import { CustomStoryObj, CustomMeta } from "test_utils/types";
 import { TaskDurationTable } from "./TaskDurationTable";
 
 export default {
   title: "Pages/Task/Table/Task Duration Table",
   component: TaskDurationTable,
-  decorators: [
-    (Story: () => JSX.Element) => (
-      <MockedProvider>
-        <Story />
-      </MockedProvider>
-    ),
-  ],
 } satisfies CustomMeta<typeof TaskDurationTable>;
 
 export const Default: CustomStoryObj<typeof TaskDurationTable> = {
